requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
dotenv==0.9.9
frozenlist==1.6.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
multidict==6.4.3
propcache==0.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-telegram-bot==22.0
setuptools==79.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
wheel==0.45.1
yarl==1.20.0
